refactor(db): move migrator to migratorce
