+++ api/bolt/migratorce/migrate_dbversion0.go
@@ -1,4 +1,4 @@
-package migrator
+package migratorce
 
 import (
 	"github.com/boltdb/bolt"
