+++ api/bolt/migratorce/migrate_dbversion11.go
@@ -1,4 +1,4 @@
-package migrator
+package migratorce
 
 import (
 	"strconv"
