+++ api/bolt/migratorce/migrate_dbversion19.go
@@ -1,4 +1,4 @@
-package migrator
+package migratorce
 
 import (
 	"strings"
