+++ api/bolt/migratorce/migrate_dbversion23.go
@@ -1,4 +1,4 @@
-package migrator
+package migratorce
 
 import portainer "github.com/portainer/portainer/api"
