+++ api/bolt/migratorce/migrate_dbversion8.go
@@ -1,15 +1,15 @@
-package migrator
+package migratorce
 
 import "github.com/portainer/portainer/api"
 
-func (m *Migrator) updateEndpointsToVersion10() error {
+func (m *Migrator) updateEndpointsToVersion9() error {
 	legacyEndpoints, err := m.endpointService.Endpoints()
 	if err != nil {
 		return err
 	}
 
 	for _, endpoint := range legacyEndpoints {
-		endpoint.Type = portainer.DockerEnvironment
+		endpoint.GroupID = portainer.EndpointGroupID(1)
 		err = m.endpointService.UpdateEndpoint(endpoint.ID, &endpoint)
 		if err != nil {
 			return err
